Reorganized, added changelog, bumped version
