+++ changelog.md
@@ -1,3 +1,13 @@
+0.16.0.0
+========
+- @bitemyapp
+  - Reorganized modules internally, ripped out Generic,
+    rewrote part of the test suite.
+- @andrewthad
+  - Added support for autogenerated elasticsearch ids in the bulk API
+- @lunaris
+  - Added support for scripts fields and function score queries
+
 0.15.0.2
 ========
 - @michaelxavier
